Missing Identifier field & use of command line certutil.exe
The Identifier (Entity ID) field is required for this to operate correctly but is not currently documented. It's set the same as Sign on URL.

Also the guide directs the user to obtain an SHA256 thumbprint of the Azure certificate by downloading OpenSSL; this is now supported by the latest version of certutil.exe built in to Windows so replaced these instructions with more straightforward ones.